+++ data-protocols/ids/ids-api-multipart-dispatcher-v1/src/main/java/org/eclipse/dataspaceconnector/ids/api/multipart/dispatcher/IdsMultipartDispatcherServiceExtension.java
@@ -1,5 +1,5 @@
 /*
- *  Copyright (c) 2021 Fraunhofer Institute for Software and Systems Engineering
+ *  Copyright (c) 2021 - 2022 Fraunhofer Institute for Software and Systems Engineering
  *
  *  This program and the accompanying materials are made available under the
  *  terms of the Apache License, Version 2.0 which is available at
@@ -9,6 +9,7 @@
  *
  *  Contributors:
  *       Fraunhofer Institute for Software and Systems Engineering - initial API and implementation
+ *       Microsoft Corporation - Use IDS Webhook address for JWT audience claim
  *
  */
 
@@ -47,10 +48,6 @@ public class IdsMultipartDispatcherServiceExtension implements ServiceExtension 
     public static final String EDC_IDS_ID = "edc.ids.id";
     public static final String DEFAULT_EDC_IDS_ID = "urn:connector:edc";
 
-    @EdcSetting
-    public static final String IDS_WEBHOOK_ADDRESS = "ids.webhook.address";
-    public static final String DEFAULT_IDS_WEBHOOK_ADDRESS = "http://localhost";
-
     @Inject
     private Monitor monitor;
     @Inject
@@ -79,10 +76,7 @@ public class IdsMultipartDispatcherServiceExtension implements ServiceExtension 
         //      once https://github.com/eclipse-dataspaceconnector/DataSpaceConnector/issues/236 is done
         var objectMapper = objectMapperFactory.getObjectMapper();
 
-        var idsWebhookAddress = getSetting(context, IDS_WEBHOOK_ADDRESS, DEFAULT_IDS_WEBHOOK_ADDRESS);
-        var idsApiPath = idsApiConfiguration.getPath();
-        var webhookPath = idsApiPath + (idsApiPath.endsWith("/") ? "data" : "/data");
-        idsWebhookAddress = idsWebhookAddress + webhookPath;
+        String idsWebhookAddress = idsApiConfiguration.getIdsWebhookAddress();
 
         var multipartDispatcher = new IdsMultipartRemoteMessageDispatcher();
         multipartDispatcher.register(new MultipartArtifactRequestSender(connectorId, httpClient, objectMapper, monitor, vault, identityService, transformerRegistry, idsWebhookAddress));