+++ spi/core-spi/src/main/java/org/eclipse/dataspaceconnector/spi/iam/TokenRepresentation.java
@@ -1,5 +1,5 @@
 /*
- *  Copyright (c) 2020, 2021 Microsoft Corporation
+ *  Copyright (c) 2020 - 2022 Microsoft Corporation
  *
  *  This program and the accompanying materials are made available under the
  *  terms of the Apache License, Version 2.0 which is available at
@@ -10,6 +10,7 @@
  *  Contributors:
  *       Microsoft Corporation - initial API and implementation
  *       Fraunhofer Institute for Software and Systems Engineering - Improvements
+ *       Microsoft Corporation - Simplified token representation
  *
  */
 
@@ -23,7 +24,6 @@ import java.util.Map;
  */
 public class TokenRepresentation {
     private String token;
-    private long expiresIn;
     private Map<String, Object> additional;
 
     private TokenRepresentation() {
@@ -35,13 +35,6 @@ public class TokenRepresentation {
      */
     public String getToken() {
         return token;
-    }
-
-    /**
-     * Returns the token expiration.
-     */
-    public long expiresIn() {
-        return expiresIn;
     }
 
     public Map<String, Object> getAdditional() {
@@ -61,11 +54,6 @@ public class TokenRepresentation {
 
         public Builder token(String token) {
             result.token = token;
-            return this;
-        }
-
-        public Builder expiresIn(long expiresIn) {
-            result.expiresIn = expiresIn;
             return this;
         }