+++ spi/core-spi/src/test/java/org/eclipse/dataspaceconnector/spi/types/domain/DataAddressTest.java
@@ -20,10 +20,8 @@ import org.junit.jupiter.api.Test;
 import java.io.IOException;
 import java.io.StringWriter;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-
 
 class DataAddressTest {
 
@@ -40,10 +38,10 @@ class DataAddressTest {
 
         DataAddress deserialized = mapper.readValue(writer.toString(), DataAddress.class);
 
-        assertNotNull(deserialized);
+        assertThat(deserialized).isNotNull();
 
-        assertEquals("test", deserialized.getType());
-        assertEquals("bar", deserialized.getProperty("foo"));
+        assertThat(deserialized.getType()).isEqualTo("test");
+        assertThat(deserialized.getProperty("foo")).isEqualTo("bar");
     }
 
     @Test
@@ -60,4 +58,15 @@ class DataAddressTest {
                 .hasMessageContaining("Property key null.");
     }
 
+    @Test
+    void verifyGetDefaultPropertyValue() {
+        assertThat(DataAddress.Builder.newInstance().type("sometype").build().getProperty("missing", "defaultValue"))
+                .isEqualTo("defaultValue");
+    }
+
+    @Test
+    void verifyGetExistingPropertyValue() {
+        assertThat(DataAddress.Builder.newInstance().type("sometype").property("existing", "existingValue").build().getProperty("existing", "defaultValue"))
+                .isEqualTo("existingValue");
+    }
 }