+++ CHANGELOG.md
@@ -21,6 +21,7 @@ in the detailed section referring to by linking pull requests or issues.
 * Event Framework for PolicyDefinition entity (#1437)
 * SQL Translation layer (#1357, #1459)
 * Permit API verbose error response (#1479)
+* Fix TODO and document `:extensions:data-plane-transfer` (#1519)
 
 #### Changed
 
@@ -29,6 +30,7 @@ in the detailed section referring to by linking pull requests or issues.
 * Refactored query capabilities for `ContractDefinition` (#1458)
 * Refactored state machine and in-memory persistence (#1511)
 * Token based Authentication can retrieve key from vault (#1537)
+* JWT audience claim check with DID (#1520)
 
 #### Removed
 
@@ -37,6 +39,7 @@ in the detailed section referring to by linking pull requests or issues.
 * Unused classes and interfaces at `ids.spi.policy` (#1471)
 * Remove modules `:extensions:transfer-functions:transfer-functions-spi` and `:extensions:transfer-functions:transfer-functions-core` (#1482)
 * Remove `ConnectorVersionProvider`, provide version as static string (#1470)
+* Remove `samples/other/run-from-junit` (#1456)
 
 #### Fixed
 
@@ -46,6 +49,7 @@ in the detailed section referring to by linking pull requests or issues.
 * Avoid endless loops in `ContractNegotiationManager` (#1487)
 * Fix race condition in `ContractNegotiationIntegrationTest` (#1505)
 * Fix for change in Cosmos DB behavior on missing sort fields (#1514)
+* Effectively removed default LIMIT in SQL Contract Def Store (#1515)
 
 ## [milestone-4] - 2022-06-07
 
@@ -376,6 +380,7 @@ in the detailed section referring to by linking pull requests or issues.
 * Add `IN` operator to all `AssetIndex` implementations (#322)
 * Support IDS logical constraint transformations (#342)
 * Add SQL persistence for contract definitions (#460) (#461)
+* Extra configuration for HttpDataSink (#1480)
 
 #### Changed